links moved to client component + css displays active state
