+++ app/page.tsx
@@ -2,7 +2,18 @@ import Input from './(client)/(widgets)/(input)/Input';
 import { getAccessToken } from '@/lib/spotify';
 
 export default async function Home() {
-  const { accessToken: token } = await getAccessToken();
+  const session = await getAccessToken();
+
+  if (!session) {
+    return (
+      <main className="flex min-h-screen flex-col justify-between p-24">
+        <p>No Session Found</p>
+      </main>
+    );
+  }
+
+  const { accessToken: token } = session;
+
   if (!token) {
     return (
       <main className="flex min-h-screen flex-col justify-between p-24">
